Configure webservers during install
We now install the config files for apache2 and lighttpd. If only one
webserver is installed, that webserver is automatically configured to
use Z-push. If more than one webserver is installed, the user is asked
for which webservers he wants Z-push enabled.
When removing Z-push, it is deconfigured from all webservers where it
was enabled.