Drop z-push-kopano metapackage
This metapackage depends on only two other package, so it has little
use. It might be confusing to have a metapackage for kopano and not
for the other backends. Also, because the dependency between z-push
and z-push-backend-kopano now runs via z-push-kopano, installing
z-push and z-push-backend-kopano together will include
z-push-backend-imap as a dependency.
All in all, this is more trouble that it's worth.