patches: Don't disable provisioning
In the latest Z-push developer meeting there was some discussion about
the fact that provisioning was by default disabled. The upstream devs
had a strong preference for keeping provisioning enabled. Since I
disabled it out of caution, not because of an actual reported issue,
let's undo that and keep provisioning enabled.